requirements.txt
streamlit
Pillow
matplotlib
numpy
